requirements.txt
discordbot.py
discord.py
ffmpeg
youtube-dl
PyNaCl
